refactor: Inline PageProgression because this component only works in this context
