+++ src/client/components/ConsentsSubHeader.tsx
@@ -1,19 +1,23 @@
 import React from 'react';
 import { css } from '@emotion/react';
-import { brand } from '@guardian/src-foundations/palette';
+import { SerializedStyles } from '@emotion/react';
+import { brand, neutral } from '@guardian/src-foundations/palette';
 import { space } from '@guardian/src-foundations';
 import { AutoRow, gridItemColumnConsents, gridRow } from '@/client/styles/Grid';
 import { from } from '@guardian/src-foundations/mq';
-import { titlepiece } from '@guardian/src-foundations/typography';
-import { PageProgression } from '@/client/components/PageProgression';
+import { textSans, titlepiece } from '@guardian/src-foundations/typography';
 import { CONSENTS_PAGES_ARR } from '@/client/models/ConsentsPages';
 import { CONSENTS_MAIN_COLOR } from '@/client/layouts/shared/Consents';
+import { SvgCheckmark } from '@guardian/src-icons';
 
 type Props = {
   autoRow: AutoRow;
   title: string;
   current?: string;
 };
+
+const BORDER_SIZE = 2;
+const CIRCLE_DIAMETER = 24;
 
 const consentsBackground = css`
   background-color: ${CONSENTS_MAIN_COLOR};
@@ -50,6 +54,74 @@ const h1 = css`
   }
 `;
 
+const circle = `
+  content: ' ';
+  box-sizing: content-box;
+  border-radius: 50%;
+  height: ${CIRCLE_DIAMETER}px;
+  position: absolute;
+  left: 0;
+  top: 0;
+  width: ${CIRCLE_DIAMETER}px;
+`;
+
+const li = (numPages: number) => css`
+  ${textSans.small()}
+  position: relative;
+  width: ${100 / numPages}%;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-end;
+  &.active {
+    ${textSans.small({ fontWeight: 'bold' })}
+  }
+  &::after,
+  &.complete::after {
+    content: ' ';
+    background-color: ${neutral[60]};
+    height: ${BORDER_SIZE}px;
+    position: absolute;
+    /* Border position from top is distance of a semicircle minus half the border thickness */
+    top: ${CIRCLE_DIAMETER / 2 + BORDER_SIZE - BORDER_SIZE / 2}px;
+    left: ${CIRCLE_DIAMETER + 2 * BORDER_SIZE}px;
+    right: 0;
+  }
+  &.complete::after {
+    height: ${BORDER_SIZE * 2}px;
+    background-color: ${brand[400]};
+  }
+  &:last-child::after {
+    display: none;
+  }
+  &::before {
+    border: ${BORDER_SIZE}px solid ${neutral[60]};
+    border-radius: 50%;
+    ${circle}
+  }
+  &.active::before,
+  &.complete::before {
+    content: ' ';
+    background-color: ${brand[400]};
+    border: ${BORDER_SIZE}px solid ${brand[400]};
+    ${circle}
+  }
+  & svg {
+    display: none;
+  }
+  &.complete svg {
+    position: absolute;
+    display: block;
+    stroke: white;
+    fill: white;
+    height: ${CIRCLE_DIAMETER}px;
+    width: ${CIRCLE_DIAMETER}px;
+    margin: ${BORDER_SIZE}px;
+    top: 0;
+    left: 0;
+    z-index: 1;
+  }
+`;
+
 const pageProgression = css`
   margin-top: ${space[5]}px;
   margin-bottom: 0;
@@ -64,6 +136,46 @@ const pageProgression = css`
     }
   }
 `;
+
+const ul = css`
+  display: flex;
+  list-style: none;
+  height: 54px;
+  padding: 0;
+  margin: 0;
+`;
+
+const PageProgression = ({
+  pages,
+  current,
+  cssOverrides,
+}: {
+  pages: string[];
+  current?: string;
+  cssOverrides?: SerializedStyles | SerializedStyles[];
+}) => {
+  const active = current ? pages.indexOf(current) : 0;
+  const getClassName = (i: number) => {
+    switch (true) {
+      case i === active:
+        return 'active';
+      case i < active:
+        return 'complete';
+      default:
+        return '';
+    }
+  };
+  return (
+    <ul css={[ul, cssOverrides]}>
+      {pages.map((page, i) => (
+        <li className={getClassName(i)} key={i} css={li(pages.length)}>
+          <SvgCheckmark />
+          <div>{page}</div>
+        </li>
+      ))}
+    </ul>
+  );
+};
 
 export const ConsentsSubHeader = ({ autoRow, title, current }: Props) => (
   <header css={[consentsBackground, ieFlexFix]}>
